+++ src/lib.rs
@@ -801,7 +801,7 @@ impl<'a> Tokenizer<'a> {
         let text = s.consume_chars(|s, c| !(c == '-' && s.starts_with(b"-->")))?;
         s.skip_string(b"-->")?;
 
-        if text.as_str().contains("--") {
+        if text.as_str().contains("--") || text.as_str().ends_with('-') {
             return Err(StreamError::UnexpectedEndOfStream); // Error type doesn't matter.
         }