+++ tests/comments.rs
@@ -5,17 +5,15 @@ use token::*;
 
 test!(comment_01, "<!--comment-->",     Token::Comment("comment", 0..14));
 test!(comment_02, "<!--<head>-->",      Token::Comment("<head>", 0..13));
-test!(comment_03, "<!----->",           Token::Comment("-", 0..8));
-test!(comment_04, "<!--<!-x-->",        Token::Comment("<!-x", 0..11));
-test!(comment_05, "<!--<!x-->",         Token::Comment("<!x", 0..10));
-test!(comment_06, "<!--<<!x-->",        Token::Comment("<<!x", 0..11));
-test!(comment_07, "<!--<<!-x-->",       Token::Comment("<<!-x", 0..12));
-test!(comment_08, "<!--<x-->",          Token::Comment("<x", 0..9));
-test!(comment_09, "<!--<>-->",          Token::Comment("<>", 0..9));
-test!(comment_10, "<!--<-->",           Token::Comment("<", 0..8));
-test!(comment_11, "<!--<--->",          Token::Comment("<-", 0..9));
-test!(comment_12, "<!--<!-->",          Token::Comment("<!", 0..9));
-test!(comment_13, "<!---->",            Token::Comment("", 0..7));
+test!(comment_03, "<!--<!-x-->",        Token::Comment("<!-x", 0..11));
+test!(comment_04, "<!--<!x-->",         Token::Comment("<!x", 0..10));
+test!(comment_05, "<!--<<!x-->",        Token::Comment("<<!x", 0..11));
+test!(comment_06, "<!--<<!-x-->",       Token::Comment("<<!-x", 0..12));
+test!(comment_07, "<!--<x-->",          Token::Comment("<x", 0..9));
+test!(comment_08, "<!--<>-->",          Token::Comment("<>", 0..9));
+test!(comment_09, "<!--<-->",           Token::Comment("<", 0..8));
+test!(comment_10, "<!--<!-->",          Token::Comment("<!", 0..9));
+test!(comment_11, "<!---->",            Token::Comment("", 0..7));
 
 macro_rules! test_err {
     ($name:ident, $text:expr) => (
@@ -59,3 +57,7 @@ test_err!(comment_err_28, "<!--<!-<!--x-->");
 test_err!(comment_err_29, "<!----!->");
 test_err!(comment_err_30, "<!----!x>");
 test_err!(comment_err_31, "<!-----x>");
+test_err!(comment_err_32, "<!----->");
+test_err!(comment_err_33, "<!------>");
+test_err!(comment_err_34, "<!-- --->");
+test_err!(comment_err_35, "<!--a--->");