+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `]]>` is no longer allowed inside a Text node.
 - Only [XML characters](https://www.w3.org/TR/xml/#char32) are allowed now.
   Otherwise, `StreamError::NonXmlChar` will occur.
+- Disallow `-` at the end of a comment. `<!--a--->` is an error now.
 
 ## [0.11.0] - 2019-11-18
 ### Added